Move Fulfillment and PurchaseDiscount componente from SearchResults to main components folder
