+++ src/components/PurchaseDiscount/PurchaseDiscount.scss
